% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cICA_morlet.R
\name{cICA_morlet}
\alias{cICA_morlet}
\title{This function performs colorICA Morlet algorithm on CIFTI files.}
\usage{
cICA_morlet(
  Xin,
  M = dim(Xin)[1],
  Win = diag(M),
  tol = 1e-04,
  maxit = 20,
  nmaxit = 5,
  dj = 0.25,
  ncores = 2
)
}
\arguments{
\item{Xin}{CIFTI file data}

\item{M}{Xin first dimension}

\item{Win}{diagonal of matrix M}

\item{tol}{tolerance}

\item{maxit}{maximum number of iteration}

\item{nmaxit}{maximum number of iteration}

\item{dj}{Divide octave in sub-octaves. If dj = 0.25 this will do 4 sub-octaves per octave.}

\item{ncores}{Number of cores to use during fitting}
}
\value{
return(as.list(result))
}
\description{
This function performs colorICA Morlet algorithm on CIFTI files.
}
