% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawplot.R
\name{drawplot}
\alias{drawplot}
\title{This function plots the signal, time-frequency power and power spectrum of Morlet wavelet transformed information from CIFTI files.}
\usage{
drawplot(obj, interpolate = TRUE)
}
\arguments{
\item{obj}{morlet transformed image data object}

\item{interpolate}{interpolation}
}
\value{
return(list(a,fig1,fig2,fig3))
}
\description{
This function plots the signal, time-frequency power and power spectrum of Morlet wavelet transformed information from CIFTI files.
}
